Add ZSink.fromFile functionality (cross compile)
